# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Central da Casa Property Finder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ImovelTipoCount {
  tipo: string;
  count: number;
}

export interface DashboardStats {
  totalClientes: number;
  totalImoveis: number;
  totalMatches: number;
  imovelPorTipo: ImovelTipoCount[];
}

export interface Cliente {
  id: number;
  nome: string;
  telefone: string;
  tipoImovelDesejado: string;
  cidade: string;
  /** @nullable */
  bairroPreferido?: string | null;
  precoMinimo: number;
  precoMaximo: number;
  quartosMinimos: number;
  /** @nullable */
  observacoes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ClienteInput {
  /** @minLength 1 */
  nome: string;
  /** @minLength 1 */
  telefone: string;
  /** @minLength 1 */
  tipoImovelDesejado: string;
  /** @minLength 1 */
  cidade: string;
  /** @nullable */
  bairroPreferido?: string | null;
  precoMinimo: number;
  precoMaximo: number;
  quartosMinimos: number;
  /** @nullable */
  observacoes?: string | null;
}

export interface ClienteList {
  data: Cliente[];
  total: number;
  page: number;
  limit: number;
}

export interface Imovel {
  id: number;
  fonte: string;
  /** @nullable */
  codigoExterno?: string | null;
  tipo: string;
  cidade: string;
  bairro: string;
  preco: number;
  area: number;
  quartos: number;
  banheiros: number;
  vagas: number;
  /** @nullable */
  descricao?: string | null;
  /** @nullable */
  urlOriginal?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ImovelList {
  data: Imovel[];
  total: number;
  page: number;
  limit: number;
}

export type MatchStatus = typeof MatchStatus[keyof typeof MatchStatus];


export const MatchStatus = {
  NOVO: 'NOVO',
  VISUALIZADO: 'VISUALIZADO',
  INTERESSADO: 'INTERESSADO',
  DESCARTADO: 'DESCARTADO',
} as const;

export interface Match {
  id: number;
  clienteId: number;
  imovelId: number;
  score: number;
  status: MatchStatus;
  createdAt: string;
}

export type MatchResultStatus = typeof MatchResultStatus[keyof typeof MatchResultStatus];


export const MatchResultStatus = {
  NOVO: 'NOVO',
  VISUALIZADO: 'VISUALIZADO',
  INTERESSADO: 'INTERESSADO',
  DESCARTADO: 'DESCARTADO',
} as const;

export interface MatchResult {
  matchId: number;
  score: number;
  status: MatchResultStatus;
  imovel: Imovel;
}

export type MatchStatusUpdateStatus = typeof MatchStatusUpdateStatus[keyof typeof MatchStatusUpdateStatus];


export const MatchStatusUpdateStatus = {
  NOVO: 'NOVO',
  VISUALIZADO: 'VISUALIZADO',
  INTERESSADO: 'INTERESSADO',
  DESCARTADO: 'DESCARTADO',
} as const;

export interface MatchStatusUpdate {
  status: MatchStatusUpdateStatus;
}

export type ListClientesParams = {
page?: number;
limit?: number;
/**
 * @nullable
 */
search?: string | null;
};

export type ListImoveisParams = {
/**
 * @nullable
 */
cidade?: string | null;
/**
 * @nullable
 */
bairro?: string | null;
/**
 * @nullable
 */
tipo?: string | null;
/**
 * @nullable
 */
precoMin?: number | null;
/**
 * @nullable
 */
precoMax?: number | null;
page?: number;
limit?: number;
};