# lib/api-zod/src/generated/types/cliente.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Central da Casa Property Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface Cliente {
  id: number;
  nome: string;
  telefone: string;
  tipoImovelDesejado: string;
  cidade: string;
  /** @nullable */
  bairroPreferido?: string | null;
  precoMinimo: number;
  precoMaximo: number;
  quartosMinimos: number;
  /** @nullable */
  observacoes?: string | null;
  createdAt: string;
  updatedAt: string;
}