# lib/api-zod/src/generated/types/clienteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Central da Casa Property Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface ClienteInput {
  /** @minLength 1 */
  nome: string;
  /** @minLength 1 */
  telefone: string;
  /** @minLength 1 */
  tipoImovelDesejado: string;
  /** @minLength 1 */
  cidade: string;
  /** @nullable */
  bairroPreferido?: string | null;
  precoMinimo: number;
  precoMaximo: number;
  quartosMinimos: number;
  /** @nullable */
  observacoes?: string | null;
}