# lib/api-zod/src/generated/types/clienteList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Central da Casa Property Finder API
 * OpenAPI spec version: 0.1.0
 */
import type { Cliente } from './cliente';

export interface ClienteList {
  data: Cliente[];
  total: number;
  page: number;
  limit: number;
}