# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Central da Casa Property Finder API
 * OpenAPI spec version: 0.1.0
 */
import type { ImovelTipoCount } from './imovelTipoCount';

export interface DashboardStats {
  totalClientes: number;
  totalImoveis: number;
  totalMatches: number;
  imovelPorTipo: ImovelTipoCount[];
}