# lib/api-zod/src/generated/types/imovel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Central da Casa Property Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface Imovel {
  id: number;
  fonte: string;
  /** @nullable */
  codigoExterno?: string | null;
  tipo: string;
  cidade: string;
  bairro: string;
  preco: number;
  area: number;
  quartos: number;
  banheiros: number;
  vagas: number;
  /** @nullable */
  descricao?: string | null;
  /** @nullable */
  urlOriginal?: string | null;
  createdAt: string;
  updatedAt: string;
}