# lib/api-zod/src/generated/types/imovelList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Central da Casa Property Finder API
 * OpenAPI spec version: 0.1.0
 */
import type { Imovel } from './imovel';

export interface ImovelList {
  data: Imovel[];
  total: number;
  page: number;
  limit: number;
}