# lib/api-zod/src/generated/types/imovelTipoCount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Central da Casa Property Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface ImovelTipoCount {
  tipo: string;
  count: number;
}