# lib/api-zod/src/generated/types/matchResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Central da Casa Property Finder API
 * OpenAPI spec version: 0.1.0
 */
import type { Imovel } from './imovel';
import type { MatchResultStatus } from './matchResultStatus';

export interface MatchResult {
  matchId: number;
  score: number;
  status: MatchResultStatus;
  imovel: Imovel;
}