# lib/api-zod/src/generated/types/matchStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Central da Casa Property Finder API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchStatusUpdateStatus } from './matchStatusUpdateStatus';

export interface MatchStatusUpdate {
  status: MatchStatusUpdateStatus;
}